refactor: split features/rbac/risk/integration.go into focused modules (Issue #454)
integration.go (1,329 lines) was a kitchen-sink handling 8+ concerns in a
single file: core integration, context building, decision enforcement, session
monitoring, continuous monitoring, risk tracking, pattern analysis, and
coordination helpers. This was inconsistent with the package's other files
(adaptive_controls.go, behavioral_analyzer.go, etc.) which are already
organized by responsibility.

Split into 6 files, each with a single cohesive responsibility:

- integration.go (~425 lines): Core RiskBasedAccessIntegration struct, config,
  ZeroTrustRiskMode constants, EnhancedRiskAccessCheck, coordinateWithZeroTrust,
  and the 4 coordination mode implementations (risk-informed, policy-informed,
  bidirectional, unified). Delegation methods for session monitoring.

- integration_context.go (162 lines): RiskContextBuilder, NewRiskContextBuilder,
  BuildRiskContext, and 5 data provider types (UserDataProvider,
  SessionDataProvider, ResourceDataProvider, EnvironmentDataProvider,
  HistoricalDataProvider) with their methods.

- integration_enforcer.go (68 lines): RiskDecisionEnforcer, NewRiskDecisionEnforcer,
  ApplyAdaptiveControls, and supporting service types (RiskSessionManager,
  RiskControlApplicator, RiskNotificationService, RiskComplianceTracker).

- integration_monitoring.go (359 lines): ContinuousRiskMonitor, SessionRiskMonitoring,
  RiskTrendPoint, RiskThresholdBreach, RiskThresholds, RiskEventCallback, RiskEvent,
  RiskEventType constants, and all monitoring lifecycle methods including
  monitoringLoop and getDefaultRiskThresholds.

- integration_tracking.go (80 lines): SessionRiskTracker, UserRiskPattern,
  RiskPatternAnalyzer, TrackRiskAssessment, GetRiskHistory.

- integration_decisions.go (279 lines): RiskPolicyAlignment constants,
  RiskFactorsSummary, coordination helper methods (convertToZeroTrustRequest,
  calculateRiskPolicyAlignment, determineRecommendedAction,
  adjustRiskScoreWithPolicyContext, calculateRiskLevelFromScore, boolToAction,
  applyCoordinationResult), and response extraction utilities (extractResourceType,
  extractRiskLevel, extractRiskScore, extractRiskFactors, abs).

Changes:
- Split integration.go from 1,329 lines into 6 focused files (1,373 total)
- All new files are under 400 lines
- No new exports, interfaces, or types — pure structural refactoring
- All tests in risk_engine_test.go pass (4 test suites, 0 failures)
- Cross-platform compilation verified (linux/amd64, linux/arm64, darwin/amd64,
  darwin/arm64, windows/amd64)
- Architecture checks pass (no central provider violations)

Fixes #454
