+++ features/rbac/risk/integration_enforcer.go
@@ -1,0 +1,68 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2026 Jordan Ritz
+package risk
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/cfgis/cfgms/api/proto/common"
+)
+
+// RiskDecisionEnforcer enforces risk-based access decisions
+type RiskDecisionEnforcer struct {
+	sessionManager      *RiskSessionManager
+	controlApplicator   *RiskControlApplicator
+	notificationService *RiskNotificationService
+	complianceTracker   *RiskComplianceTracker
+}
+
+// NewRiskDecisionEnforcer creates a new risk decision enforcer
+func NewRiskDecisionEnforcer() *RiskDecisionEnforcer {
+	return &RiskDecisionEnforcer{
+		sessionManager:      &RiskSessionManager{},
+		controlApplicator:   &RiskControlApplicator{},
+		notificationService: &RiskNotificationService{},
+		complianceTracker:   &RiskComplianceTracker{},
+	}
+}
+
+// ApplyAdaptiveControls applies the determined adaptive controls
+func (rde *RiskDecisionEnforcer) ApplyAdaptiveControls(ctx context.Context, request *common.AccessRequest, controls []AdaptiveControl) error {
+	for _, control := range controls {
+		err := rde.controlApplicator.ApplyControl(ctx, request, &control)
+		if err != nil {
+			return fmt.Errorf("failed to apply control %s: %w", control.Type, err)
+		}
+	}
+
+	// Track compliance with applied controls
+	err := rde.complianceTracker.TrackControlApplication(ctx, request, controls)
+	if err != nil {
+		// Log but don't fail
+		fmt.Printf("Warning: Failed to track compliance: %v", err)
+	}
+
+	return nil
+}
+
+// Supporting service types
+
+type RiskSessionManager struct{}
+
+type RiskControlApplicator struct{}
+
+func (rca *RiskControlApplicator) ApplyControl(ctx context.Context, request *common.AccessRequest, control *AdaptiveControl) error {
+	// Simplified control application - in practice would interact with session management, monitoring, etc.
+	fmt.Printf("Applied control: %s with parameters: %+v", control.Type, control.Parameters)
+	return nil
+}
+
+type RiskNotificationService struct{}
+
+type RiskComplianceTracker struct{}
+
+func (rct *RiskComplianceTracker) TrackControlApplication(ctx context.Context, request *common.AccessRequest, controls []AdaptiveControl) error {
+	// Simplified compliance tracking
+	return nil
+}
